Refactored todo list omponent
