+++ Week-12/todo/src/App.js
@@ -3,7 +3,6 @@ import "./App.css";
 import { Route, Routes } from "react-router-dom";
 
 import Home from "./pages/Home";
-import Hooks from "./pages/Hooks";
 import Navbar from "./components/navbar/navbar";
 import Todo from "./pages/Todo.jsx";
 
@@ -16,8 +15,6 @@ function App() {
         <div className="layout_container">
           <Routes>
             <Route path="/" element={<Home />} />
-            {/* custom hooks practice */}
-            <Route path="/hooks" element={<Hooks />} />
             <Route path="/todo" element={<Todo />} />
           </Routes>
         </div>
